refactor: amalgamate file related functions to single file
