--- packages/lpa-questionnaire-web-app/src/lib/mime-types.js
@@ -1,8 +1,0 @@
-module.exports = {
-  MIME_TYPE_DOC: 'application/msword',
-  MIME_TYPE_DOCX: 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-  MIME_TYPE_PDF: 'application/pdf',
-  MIME_TYPE_TIF: 'image/tiff',
-  MIME_TYPE_JPEG: 'image/jpeg',
-  MIME_TYPE_PNG: 'image/png',
-};
