+++ packages/lpa-questionnaire-web-app/src/validators/custom/file-size.js
@@ -1,4 +1,4 @@
-const fileSizeDisplayHelper = require('../../lib/file-size-display-helper');
+const { fileSizeDisplayHelper } = require('../../lib/file-upload-helpers');
 
 module.exports = (givenFileSize, maxFileSize, fileName = 'The file') => {
   if (givenFileSize > maxFileSize) {
